% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWASbyGLM.R
\name{GWASbyGLM}
\alias{GWASbyGLM}
\title{GWASbyGLM}
\usage{
GWASbyGLM(mdpN, mdpPheno, mdpCov, option = "cov")
}
\arguments{
\item{mdpN}{an input numeric genotype matrix coded as c(0,1,2). Rows are samples, columns are markers.}

\item{mdpPheno}{an input phenotype matrix. Rows are samples, columns are traits.}

\item{option}{what covariates to include in the model as fixed effects. 'cov' for user defined covariates, 'pca' for principle component inclusion.}
}
\value{
A vector of p-values corresponding to each marker.
}
\description{
This function performs a genome-wide-association study using the generalized linear model paradigm. Users have the option to include their own covariates, but can also perform principle component analysis.
If so, the 'princomp' function is applied to the genotype matrix and components that explain more than 1 percent of the variance are included as long as they are not also correlated with user provided covariates.
}
\examples{
# Reading example data from the Github repository
mdpN <-read.csv(text=getURL("https://raw.githubusercontent.com/mtmcgowan/CROPS545_HW4_GLM_GWAS/master/example_mdpN"), header=T, sep = " ")
mdpPheno <-read.csv(text=getURL("https://raw.githubusercontent.com/mtmcgowan/CROPS545_HW4_GLM_GWAS/master/example_mdpPheno"), header=T, sep = " ")
mdpCov <-read.csv(text=getURL("https://raw.githubusercontent.com/mtmcgowan/CROPS545_HW4_GLM_GWAS/master/example_mdpCov"), header=T, sep = " ")

# Running the function only including covariates as fixed effects
test_results <- GWASbyGLM(mdpN, mdpPheno, mdpCov, option='cov')
}
