% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manhattan.R
\name{plot_manhattan}
\alias{plot_manhattan}
\title{plot_manhattan}
\usage{
plot_manhattan(mdpSNPinfo, GWASout, error = 0.05)
}
\arguments{
\item{mdpSNPinfo}{a data.frame with 3 columns containing snp name, chromosome, and basepair position. SNPs must be in the same order as the mdpN genotype matrix used for GWASbyGLM}

\item{GWASout}{a vector containing p-values generated with the GWASbyGLM function}

\item{error}{a value from 0 to 1 indicating what type-1 error threshold should be used for calculating a Bonferroni cutoff. Default set to 0.05.}
}
\value{
a manhattan plot with a horizontal line representing the Bonferroni correction cutoff.
}
\description{
This function performs a genome-wide-association study using the generalized linear model paradigm. Users have the option to include their own covariates, but can also perform principle component analysis.
If so, the 'princomp' function is applied to the genotype matrix and components that explain more than 1% of the variance are included as long as they are not also correlated with user provided covariates.
}
\examples{
# Reading example data from the Github repository
GWASout <-read.csv(text=getURL("https://raw.githubusercontent.com/mtmcgowan/CROPS545_HW4_GLM_GWAS/master/example_GWASout"), header=T, sep = " ", row.names = 1)
mdpSNPinfo <-read.csv(text=getURL("https://raw.githubusercontent.com/mtmcgowan/CROPS545_HW4_GLM_GWAS/master/example_mdpSNPinfo"), header=T, sep = " ")

# Generating the manhattan plot
plot_manhattan(mdpSNPinfo, GWASout)
}
