% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manhattan.R
\name{plot_manhattan}
\alias{plot_manhattan}
\title{plot_manhattan}
\usage{
plot_manhattan(mdpSNPinfo, GWASout, error = 0.05)
}
\arguments{
\item{mdpSNPinfo}{a data.frame with 3 columns containing snp name, chromosome, and basepair position. SNPs must be in the same order as the mdpN genotype matrix used for GWASbyGLM}

\item{GWASout}{a vector containing p-values generated with the GWASbyGLM function}

\item{error}{a value from 0 to 1 indicating what type-1 error threshold should be used for calculating a Bonferroni cutoff. Default set to 0.05.}
}
\value{
a manhattan plot with a horizontal line representing the Bonferroni correction cutoff.
}
\description{
This function takes the p-value output from GWASbyGLM and combines it with SNP positional information to generate a manhattan plot using the 'qqman' package
}
\examples{
# Reading example data from the Github repository
GWASout <-read.csv(text=getURL("https://raw.githubusercontent.com/mtmcgowan/CROPS545_HW4_GLM_GWAS/master/example_GWASout"), header=T, sep = " ", row.names = 1)
mdpSNPinfo <-read.csv(text=getURL("https://raw.githubusercontent.com/mtmcgowan/CROPS545_HW4_GLM_GWAS/master/example_mdpSNPinfo"), header=T, sep = " ")

# Generating the manhattan plot
plot_manhattan(mdpSNPinfo, GWASout)
}
